% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{Construct technical matrices}
\usage{
construct()
}
\value{
WIOT A list containig input-output table and other parameters. Function adds/updates
  Leontieff Inverse L, technical coefficient matrix A, Final Demand matrix, value added per
  unit of gross output vector.
}
\description{
Construct technical matrices
}
\examples{
get_io(year = 2009, version = "B")
construct()

get_io(2008)
construct()

}
