% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdvac.R
\name{fdvac}
\alias{fdvac}
\title{Final demand decomposition for selected country}
\usage{
fdvac(country = 10, calc = TRUE)
}
\arguments{
\item{country}{A number 1:41 for version A, 1:44 for version B, see WIOT$countries.}

\item{calc}{A boolean, for calc = TRUE result as a share of total final demand, for
calc = FALSE result in mln USD.}
}
\value{
A list contains title, country, year and the final demand decomposition.
}
\description{
Final demand decomposition for selected country
}
\examples{
get_io(year = 1998, version = "A")
construct()
fdvac(10)

fdvac(22, calc = FALSE)

}
