% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdvad.R
\name{fdvad}
\alias{fdvad}
\title{Final demand decomposition for selected country and sector}
\usage{
fdvad(country = 10, sector = 15)
}
\arguments{
\item{country}{A number 1:41 for version A, 1:44 for version B, see WIOT$countries.}

\item{sector}{A number 1:35 for version A, 1:56 for version B, see WIOT$sectors.}
}
\value{
A list contains title, country, year, sector and the final demand decomposition.
}
\description{
Final demand decomposition for selected country and sector
}
\examples{
get_io(year = 2004, version = "B")
construct()
fdvad(34,12)

fdvad(12,11)

}
