% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_io.R
\name{get_io}
\alias{get_io}
\title{Get input-output data}
\usage{
get_io(year = 1995, version = "A")
}
\arguments{
\item{year}{A number indicating the year of the input-output table to be downloaded. There are
17 distinct datasets for version A (1995-2011) and 15 for version B (2000-2014).}

\item{version}{A character ("A", or "B") indicating the release version of the input-output table
to be downloaded.}
}
\value{
WIOT A list containig input-output table and other parameters, which is an input
   for construct() function.
}
\description{
Get input-output data
}
\examples{
get_io(year = 2014, version = "B")
get_io(2008)
}
