% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstr.R
\name{upstr}
\alias{upstr}
\title{Level of upstreamness}
\usage{
upstr(country = NA, sector = NA)
}
\arguments{
\item{country}{A number 1:41 for version A, 1:44 for version B, see WIOT$countries.}

\item{sector}{A number 1:35 for version A, 1:56 for version B, see WIOT$sectors.}
}
\value{
Level of upstreamness for selected country and sector / all sectors in the country
        / the same sector in all countries / all sectors in all countries.
}
\description{
Level of upstreamness
}
\examples{
get_io(year = 2004, version = "B")
construct()
upstr(34,12)

upstr()

upstr(,45)

}
